requirements.txt
fastapi
uvicorn
psycopg2-binary
pyyaml
openai
qdrant-client
pypdf2
tiktoken